@@ -10,7 +10,7 @@
         }
     ],
     "require": {
-        "php": ">=7.0.0",
+        "php": ">=7.2.0",
         "illuminate/console": "^6.0",
         "illuminate/contracts": "^6.0",
         "illuminate/support": "^6.0",
